feat(appium): cull invalid data from manifest
If, for some reason, an extension doesn't exist on disk yet does exist in the manifest, it probably shouldn't be in the manifest, right? Anyway:

This change causes invalid entries to be culled in the same operation as "sync with installed packages". Those situations are:

- if the manifest needs a migration
- if `appium` is a local dep and its package.json has changed

An argument could be made that we want to cull invalid entries regardless of whether we're using a local `appium` or `APPIUM_HOME`, and likewise on any extension subcommand which can modify the manifest (install, uninstall, update). While good housekeeping is in order, I'm not sure the overhead is worth it? I don't know.